Support EPUB generation of Consensus.
The biggest issue was that, while EPUB is nominally supported as a
Pandoc output type, it only exists as a format to "match" on. You
still have to use "html" as the output format. This means that all
functions which called `pandoc.RawBlock(format, "...")` fail to output
in EPUB mode.

This has theoretically been fixed. There still may be challenges,
including, it's unclear whether epub readers handle more advanced CSS
like `content` styles in `::before` rules, so we might need a
different method for outputting the voices in consensus.epub anyway.